apps/food-suggestion-ai-backend/pyproject.toml
[project]
name = "food-suggestion-ai-backend"
version = "0.1.0"
description = "AI-powered food suggestion backend using FastAPI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.3",
    "pydantic-settings>=2.6.1",
    "openai>=1.55.3",
    "google-generativeai>=0.8.3",
    "httpx>=0.27.2",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
